call_notes_app/agentcore_agent/research_agent/requirements.txt
bedrock-agentcore
strands-agents
boto3>=1.35.0
duckduckgo-search>=6.0.0
